implementationb

+++ backend/src/modules/lesson/lesson.service.ts
@@ -1,8 +1,7 @@
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
-import { CreateLessonDto } from './dto/create-lesson.dto';
 import { addDays, endOfMonth, parseISO, getDay, startOfDay, format } from 'date-fns';
 import { Cron, CronExpression } from '@nestjs/schedule';
-import { CancelLessonDto } from './dto/cancel-lesson.dto';
+import { CancelationStatusEnum, CancelLessonDto } from './dto/cancel-lesson.dto';
 import { SingleLessonInputDto } from './dto/single-lesson.input.dto';
 import { LessonRepository } from './lesson.repository';
 import { PlanService } from '../plan/plan.service';
@@ -11,35 +10,71 @@ import { RegularLessonsInputDto, WeekDay } from './dto/regular-lesson.input.dto'
 import { LessonRegularRepository } from './lesson-regular.repository';
 import { RegularLessonOutputDto } from './dto/regular-lesson.output.dto';
 import { PlanTypeEnum } from '../plan/dto/create-plan.input.dto';
-
+import { StudentService } from '../student/student.service';
+import { LessonInputStatusEnum, LessonStatusEnum } from './dto/lesson-status.enum';
+import { Lesson } from '@prisma/client';
+import { ChangeTeacherDto } from './dto/change-teacher.dto';
+import { TeacherService } from '../teacher/teacher.service';
+import { JwtPayloadDto } from '../auth/dto/jwt.payload.dto';
+import { TeacherRoleEnum } from '../teacher/dto/teacherRole';
 
 @Injectable()
 export class LessonService {
 	constructor(
 		private readonly lessonRepository: LessonRepository,
 		private readonly planService: PlanService,
+		private readonly studentService: StudentService,
+		private readonly teacherService: TeacherService,
 		private readonly lessonRegularRepository: LessonRegularRepository,
 	) { }
 
-	// async createSingleLesson(singleLessonInputDto: SingleLessonInputDto): Promise<LessonOutputDto> {
-	// 	const { plan_id, start_date, student_id, teacher_id, corrected_time, status, rescheduled_lesson_id, rescheduled_lesson_date } = singleLessonInputDto;
+	async createSingleLessonByAdmin(singleLessonInputDto: SingleLessonInputDto): Promise<LessonOutputDto> {
+		const { plan_id, start_date, student_id, teacher_id, isFree } = singleLessonInputDto;
 
-	// 	const lessonAlreadyBooked = await this.lessonRepository.getLessonByStartDateAndStudentId(start_date, student_id);
+		const date = new Date(start_date);
+		const plan = await this.planService.findById(plan_id);
+		if (!plan) {
+			throw new NotFoundException('Plan not found');
+		}
+		if (plan.deleted_at) {
+			throw new BadRequestException('Plan already deleted');
+		}
+		const student = await this.studentService.findById(student_id);
+		if (!student) {
+			throw new NotFoundException('Student not found');
+		}
 
-	// 	if (lessonAlreadyBooked) {
-	// 		throw new BadRequestException("Lesson already booked for this student at this time");
-	// 	}
-
-	// 	const plan = await this.planService.findById(plan_id);
-	// 	if (!plan) {
-	// 		throw new NotFoundException('Plan not found');
-	// 	}
-
-	// 	return await this.lessonRepository.createLesson(singleLessonInputDto);
-	// }
+		const lessonAlreadyBooked = await this.lessonRepository.findExistingLessonsByDate(date);
+		if (lessonAlreadyBooked.length > 1) {
+			throw new BadRequestException(`Максимальное количество уроков в это время: ${date.toLocaleDateString()}`);
+		}
+		if (lessonAlreadyBooked.length === 1 && lessonAlreadyBooked[0].plan.plan_type === PlanTypeEnum.INDIVIDUAL) {
+			throw new BadRequestException(`Это время занято индивидуальным занятием у ${lessonAlreadyBooked[0].student.name}: ${date.toLocaleDateString()}`);
+		}
+		if (lessonAlreadyBooked.filter(el => el.student.id === student_id).length > 0) {
+			throw new BadRequestException(`Это время уже назначено у ${lessonAlreadyBooked[0].student.name}: ${date.toLocaleDateString()}`);
+		}
+		if (lessonAlreadyBooked.length > 0 && lessonAlreadyBooked[0].plan_id !== plan_id) {
+			throw new BadRequestException(`Не совпадает тарифный план: ${date.toLocaleDateString()}`);
+		}
+		const newLesson = {
+			student_id,
+			teacher_id,
+			plan_id,
+			date,
+			is_free: isFree,
+			is_regular: false,
+			status: LessonInputStatusEnum.PENDING_UNPAID,
+		}
+		return await this.lessonRepository.createSingleLesson(newLesson as Lesson);
+	}
 
 	async findLessonsForPeriod(start_date: string, end_date: string, teacher_id: number): Promise<LessonOutputDto[]> {
 		return await this.lessonRepository.findLessonsForPeriod(start_date, end_date, teacher_id);
+	}
+
+	async findLessonsByStartDate(start_date: Date, teacher_id: number): Promise<LessonOutputDto[]> {
+		return await this.lessonRepository.findLessonsByStartDate(start_date, teacher_id);
 	}
 
 	async createRegularLessons(regularLessonsInputDto: RegularLessonsInputDto, student_id: number): Promise<RegularLessonOutputDto[]> {
@@ -90,7 +125,7 @@ export class LessonService {
 					await this.lessonRegularRepository.deleteRegularLesson(regularLesson.id);
 					throw new BadRequestException(`Это время уже назначено у ${existingLessons[0].student.name}: ${mergedDate}`);
 				}
-				if (existingLessons[0].plan_id !== plan_id) {
+				if (existingLessons.length > 0 && existingLessons[0].plan_id !== plan_id) {
 					await this.lessonRegularRepository.deleteRegularLesson(regularLesson.id);
 					throw new BadRequestException(`Не совпадает тарифный план: ${mergedDate}`);
 				}
@@ -100,6 +135,39 @@ export class LessonService {
 		}
 		return regularLessons;
 	}
+
+	async changeTeacher(lessonId: number, changeTeacherDto: ChangeTeacherDto): Promise<void> {
+		const teacher = await this.teacherService.getTeacherById(changeTeacherDto.teacher_id);
+		if (!teacher) {
+			throw new NotFoundException('Teacher not found');
+		}
+		if (teacher.deleted_at) {
+			throw new BadRequestException('Teacher is deleted');
+		}
+		await this.lessonRepository.changeTeacher(lessonId, changeTeacherDto.teacher_id);
+	}
+
+	async cancelLesson(lessonId: number, cancelLessonDto: CancelLessonDto, teacher: JwtPayloadDto): Promise<void> {
+		const lesson = await this.lessonRepository.findById(lessonId);
+		if (!lesson) {
+			throw new NotFoundException('Lesson not found');
+		}
+		if (lesson.status === LessonStatusEnum.CANCELLED || lesson.status === LessonStatusEnum.MISSED || lesson.status === LessonStatusEnum.RESCHEDULED) {
+			throw new BadRequestException('Lesson already cancelled');
+		}
+		const existingTeacher = await this.teacherService.getTeacherById(+teacher.id);
+		if (!existingTeacher) {
+			throw new NotFoundException('Teacher not found');
+		}
+		if (existingTeacher.deleted_at) {
+			throw new BadRequestException('Teacher is deleted');
+		}
+		if (teacher.role !== TeacherRoleEnum.ADMIN && lesson.student.teacher_id !== +teacher.id) {
+			throw new BadRequestException('You are not allowed to cancel this lesson');
+		}
+		await this.lessonRepository.cancelLesson(lessonId, cancelLessonDto);
+	}
+
 
 	private getDatesForWeekDay(weekDay: WeekDay, startDate: string, endDate: string): Date[] {
 		// Parse ISO dates (already in UTC)
@@ -156,295 +224,6 @@ export class LessonService {
 		};
 		return mapping[weekDay];
 	}
-
-	// async create(createLessonDto: CreateLessonDto) {
-	// 	const { bookUntilCancellation, plan_id, specificDays, start_date, student_id, corrected_time } = createLessonDto;
-
-	// 	const corrected_time_date = new Date(corrected_time);
-	// 	const lessonAlreadyBooked = await this.prisma.lesson.findFirst({
-	// 		where: {
-	// 			start_date,
-	// 			student_id,
-	// 		},
-	// 	});
-
-	// 	if (lessonAlreadyBooked) {
-	// 		throw new BadRequestException("Lesson already booked for this student");
-	// 	}
-
-
-	// 	const plan = await this.prisma.plan.findUnique({
-	// 		where: {
-	// 			id: plan_id,
-	// 		},
-	// 	});
-
-	// 	// проверяем, нет ли уже занятия на эту дату
-	// 	await this.checkDaysForLessons(start_date, plan);
-
-	// 	if (bookUntilCancellation) {
-	// 		await this.prisma.student.update({
-	// 			data: {
-	// 				bookUntilCancellation: true,
-	// 			},
-	// 			where: {
-	// 				id: student_id,
-	// 			},
-	// 		});
-
-	// 		const start = new Date(start_date);
-	// 		const end = endOfMonth(start);
-	// 		const weekday = start.getDay(); // 0 = Sunday, 1 = Monday, ..., 6 = Saturday
-
-	// 		const lessonDates: Date[] = [];
-
-	// 		// Начинаем с даты старта и двигаемся по дням
-	// 		let current = new Date(start);
-	// 		while (current <= end) {
-	// 			if (current.getDay() === weekday) {
-	// 				lessonDates.push(new Date(current));
-	// 			}
-	// 			current = addDays(current, 1);
-	// 		}
-
-	// 		const daysIsFree = await this.daysIsFree(lessonDates as any, plan);
-
-	// 		if (daysIsFree) {
-	// 			// Создаём занятия на каждый найденный день
-	// 			for (const date of lessonDates) {
-
-	// 				const merged = new Date(date);
-	// 				merged.setUTCHours(corrected_time_date.getUTCHours(), corrected_time_date.getUTCMinutes(), corrected_time_date.getUTCSeconds(), corrected_time_date.getUTCMilliseconds());
-	// 				await this.prisma.lesson.create({
-	// 					data: {
-	// 						plan_id,
-	// 						start_date: date,
-	// 						student_id,
-	// 						corrected_time: merged.toISOString(),
-	// 						is_regular: true
-	// 					},
-	// 				});
-	// 			}
-
-	// 			return {
-	// 				message: `Lessons booked every ${start.toLocaleDateString('en-US', { weekday: 'long' })} until end of month`,
-	// 			};
-	// 		}
-	// 	} else if (specificDays.length > 0) {
-	// 		const updatedDates = this.updateSpecificDays(specificDays, start_date);
-
-	// 		const daysIsFree = await this.daysIsFree(updatedDates as any, plan);
-
-	// 		if (daysIsFree) {
-	// 			updatedDates.forEach(async (day) => {
-	// 				const merged = new Date(day);
-	// 				merged.setUTCHours(corrected_time_date.getUTCHours(), corrected_time_date.getUTCMinutes(), corrected_time_date.getUTCSeconds(), corrected_time_date.getUTCMilliseconds());
-	// 				await this.prisma.lesson.create({
-	// 					data: {
-	// 						plan_id,
-	// 						start_date: day,
-	// 						student_id,
-	// 						corrected_time: merged.toISOString(),
-	// 						is_regular: true
-	// 					},
-	// 				});
-	// 			});
-	// 		}
-
-	// 		return {
-	// 			message: `Lessons booked on specific days`,
-	// 		};
-	// 	}
-
-
-
-	// 	const newLesson = await this.prisma.lesson.create({
-	// 		data: {
-	// 			plan_id,
-	// 			start_date,
-	// 			student_id,
-	// 			corrected_time,
-	// 			rescheduled_lesson_id: createLessonDto.rescheduled_lesson_id || null,
-	// 			rescheduled_lesson_date: createLessonDto.rescheduled_lesson_date || null,
-	// 			is_regular: false
-	// 		},
-	// 	});
-
-	// 	if (createLessonDto.rescheduled_lesson_id) {
-	// 		await this.prisma.lesson.update({
-	// 			data: {
-	// 				rescheduled_to_lesson_id: newLesson.id,
-	// 				rescheduled_to_lesson_date: newLesson.corrected_time,
-	// 			},
-	// 			where: {
-	// 				id: createLessonDto.rescheduled_lesson_id,
-	// 			},
-	// 		});
-	// 	}
-
-	// 	return {
-	// 		message: "Lesson created successfully",
-	// 	};
-	// }
-
-	// async findLessonsForPeriod(start_date: string, end_date: string) {
-	// 	return await this.prisma.lesson.findMany({
-	// 		include: {
-	// 			student: true,
-	// 			plan: true,
-	// 		},
-	// 		where: {
-	// 			start_date: {
-	// 				gte: start_date,
-	// 				lte: end_date,
-	// 			},
-	// 		},
-	// 	});
-	// }
-
-	// async cancelLesson(id: number, cancelLessonDto: CancelLessonDto) {
-	// 	return await this.prisma.lesson.update({
-	// 		data: {
-	// 			status: cancelLessonDto.cancelationType,
-	// 			comment: cancelLessonDto.comment,
-	// 		},
-	// 		where: { id },
-	// 	});
-	// }
-
-	// findOne(id: number) {
-	// 	return `This action returns a #${id} lesson`;
-	// }
-
-	// update(id: number) {
-	// 	return `This action updates a #${id} lesson`;
-	// }
-
-	// remove(id: number) {
-	// 	return `This action removes a #${id} lesson`;
-	// }
-
-	// async findLessonsByStartDate(start_date: string) {
-	// 	return await this.prisma.lesson.findMany({
-	// 		where: {
-	// 			start_date,
-	// 		},
-	// 		include: {
-	// 			student: true,
-	// 			plan: true,
-	// 		},
-	// 	});
-	// }
-
-	// private updateSpecificDays(specificDays: string[], start_date: string) {
-	// 	const start_date_date = new Date(start_date);
-	// 	const hours = start_date_date.getHours();
-	// 	const minutes = start_date_date.getMinutes();
-	// 	const seconds = start_date_date.getSeconds();
-	// 	const ms = start_date_date.getMilliseconds();
-
-	// 	const updatedDates = specificDays.map(dateStr => {
-	// 		const original = new Date(dateStr);
-
-	// 		// Создаём новую дату с той же датой, но с временем из start_date
-	// 		const updated = new Date(
-	// 			original.getFullYear(),
-	// 			original.getMonth(),
-	// 			original.getDate(),
-	// 			hours,
-	// 			minutes,
-	// 			seconds,
-	// 			ms
-	// 		);
-
-	// 		return updated.toISOString(); // 👉 формат: 2025-09-09T08:00:00.000Z
-	// 	});
-
-	// 	return updatedDates;
-	// }
-
-	// private async checkDaysForLessons(start_date: string, plan: any): Promise<boolean> {
-	// 	const lessons = await this.prisma.lesson.findMany({
-	// 		where: {
-	// 			start_date,
-	// 			status: {
-	// 				not: "CANCELLED",
-	// 			},
-	// 		},
-	// 		include: {
-	// 			student: true,
-	// 			plan: true,
-	// 		},
-	// 	});
-
-	// 	if (lessons.length === 0) {
-	// 		return true;
-	// 	}
-
-	// 	if (lessons[0].plan.plan_type === PlanType.INDIVIDUAL && lessons.length >= 1) {
-	// 		const message = `Lesson already exists: ${lessons[0].start_date.toLocaleDateString()} ${lessons[0].plan.plan_name} - ${lessons[0].student.name}`;
-	// 		throw new BadRequestException(message);
-	// 	}
-
-	// 	if (lessons[0].plan.plan_type === PlanType.PAIR && (lessons.length >= 2)) {
-	// 		const message = `Lesson already exists: ${lessons[0].start_date.toLocaleDateString()} ${lessons[0].plan.plan_name} - ${lessons[0].student.name} and ${lessons[1].student.name}`;
-	// 		throw new BadRequestException(message);
-	// 	}
-
-	// 	if (lessons[0].plan.plan_type === PlanType.PAIR && lessons.length === 1 && plan?.plan_type === PlanType.INDIVIDUAL) {
-	// 		const message = `You cant assign individual plan to pair lesson: ${lessons[0].start_date.toLocaleDateString()} ${lessons[0].plan.plan_name} - ${lessons[0].student.name}`;
-	// 		throw new BadRequestException(message);
-	// 	}
-
-	// 	if (lessons[0].plan.plan_type === PlanType.PAIR && lessons.length === 1 && plan?.plan_type === PlanType.PAIR && lessons[0].plan.duration !== plan.duration) {
-	// 		const message = `You cant assign pair plan with different duration to pair lesson: ${lessons[0].start_date.toLocaleDateString()} ${lessons[0].plan.plan_name} - ${lessons[0].student.name}`;
-	// 		throw new BadRequestException(message);
-	// 	}
-
-
-	// 	return true;
-	// }
-
-	// private async daysIsFree(days: Date[], plan: any): Promise<boolean> {
-	// 	const busyDays = await this.prisma.lesson.findMany({
-	// 		where: {
-	// 			start_date: {
-	// 				in: days,
-	// 			},
-	// 		},
-	// 		include: {
-	// 			plan: true,
-	// 			student: true,
-	// 		},
-	// 	});
-
-	// 	let unavailableBusyDays: any[] = [];
-	// 	if (plan.plan_type === PlanType.PAIR) {
-	// 		const notPairOrDuration = busyDays.filter(lesson => lesson.plan.plan_type !== PlanType.PAIR || lesson.plan.duration !== plan.duration);
-	// 		const busyPairs = busyDays.filter(lesson => lesson.plan.plan_type === PlanType.PAIR && lesson.plan.duration === plan.duration);
-	// 		// Группируем по start_date
-	// 		const groupedByDate = busyPairs.reduce((acc, lesson) => {
-	// 			const date = lesson.start_date.toISOString();
-	// 			acc[date] = acc[date] ? [...acc[date], lesson] : [lesson];
-	// 			return acc;
-	// 		}, {} as Record<string, typeof busyPairs>);
-
-	// 		// Оставляем только те даты, где больше одной записи
-	// 		const duplicates = Object.values(groupedByDate)
-	// 			.filter(group => group.length > 1)
-	// 			.flat(); // объединяем в один массив
-	// 		unavailableBusyDays = [...notPairOrDuration, ...duplicates];
-	// 	} else {
-	// 		unavailableBusyDays = busyDays;
-	// 	}
-
-	// 	if (unavailableBusyDays.length > 0) {
-	// 		const message = `Lesson already exists: ${busyDays[0].start_date.toLocaleDateString('ru-RU', { day: '2-digit', month: '2-digit', year: 'numeric' })} ${busyDays[0].plan.plan_name} - ${busyDays[0].student.name}`;
-	// 		throw new BadRequestException(message);
-	// 	}
-
-	// 	return true;
-	// }
 
 	@Cron(CronExpression.EVERY_DAY_AT_MIDNIGHT)
 	async updateLessonsStatus() {
